pogo_automation/requirements.txt
selenium
pytest
pytest-xdist
pytest-html
python-dotenv
pytest-ordering
pytest_html
